backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-extra-fields==3.0.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
PySocks==1.7.1
pytest==7.2.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tldextract==3.4.0
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
progress==1.6
